Support for report scheduling and exporting in a personal devel cluster.
- Adds a new Kubernetes secret and corresponding volume mount for
storing the private key json file for a service account that allows
the ReportMaster to access a Google Cloud Storage bucket.

- Adds new command .cobaltb.py deploy upload_service_account_key

- Adds new flag --enable_report_scheduling

- updates the README with a description of how to create a new
GCS bucket and a service account key to access it. (I also
took the opportunity to alphebatize the description of the
cluster.json file.)

- Changes the name of the environment variable used to refer to the
path to the secret key json file. Previously we were using the
generic GOOGLE_APPLICATION_CREDENTIALS and we here change to
a more specific COBALT_GCS_ACCOUNT_CREDENTIALS.

- Reorganized the Dockerfile for ReportMaster. We moved the
"apt-get update" and "apt-get install" for git up above the
COPY analyzer_report_master statement. This makes a huge
difference in the amuount of time it takes to build and
upload the image after a code change.

Change-Id: Id455e95acd6c09f1bdce7037a8d0f4b86a9e8214