Fix handling of the flag --enable_report_scheduling for ReportMaster.
There were several bugs which together resulted in the behavior
that the flag was always enabled. This CL fixes those bugs
so that, as desired, the feature is disabled by default.

We also fixed an unrelated problem in container_util.py: The
command gcloud service-managment has been deprecated and so
we were seeing warnings about that. Its replacement is
"gcloud endpoints services"

Change-Id: Ia285dfe6125fb46bd423f7658d758e03ef983a3e