Reworked the handling of file paths used when exporting to GCS.
- Previously we provided a field in the ExportConfig called folder_path
where a folder path could be specified. But in fact we really wanted
this to always be of a certain form and it was error prone so
now we have removed that field and we compute it programatically instead.

- Previously we named the exported CSV files using their integer day
index but it is much more user friendly to name the file using a
natural Calendar Date syntax so we switched to that.

Change-Id: Ib30c19c5402d54ac43851505ba1b82ddde91b8de