Disable flaky test case EnableReportScheduling
Our tests in CI and CQ have become too flaky so I am disabling this
one test case that fails about 20% of the time.

EnableReportScheduling will not execute unless the flag
--tests=btemulator is passed meaning that only the bt emulator
tests should be run. In this case we run the flaky test.
Since this flag is not being passed on the bots it should
remove the flakiness on the bots.

Change-Id: I9167824b7e47ff3b292f833f94c7e7c0b9d46f79