ReportExporter: Use the stream version of GcsUtil::Upload().
We change ReportExporter to use the new stream version of the Upload()
method.

But in this CL we don't yet change ReportSerializer to serialize to
a stream. ReportSerializer still serializes to a string and so
in ReportExporter for now we just wrap that string with a stream.

In a follow-up CL we will change ReportSerializer to serialize to
a stream instead of a string.

Change-Id: I569a619ca75279d70fbc1030a522e2ab2d52bb4d