ReportSerializer: Adds functionality for incremental serialization.
This is in preparation for being able to serialize reports that don't
fit in memory. We add the methods StartSerializingReport() and
AppendRows() to ReportSerializer. These methods serialize to an output
stream instead of to a string.

As part of this we refactor the class ReportSerializer so that it
has state: We move some of the arguments that used to be in
SerializeReport() to the constructor.

Change-Id: I9574713330dc5a4ad014fb20388eebd3d8d508bb