Adds ReportStream
A ReportStream is an input stream from which one may read a serialized
report. A ReportStream contains a ReportSerializer and a ReportRowIterator.
As more bytes are read from the ReportStream, more rows are read from the
ReportRowIterator and serialized using the ReportSerializer. In this way it
is possible to serialize a report without having the whole report in memory
at once.

Change-Id: Ie528f5a0ed3cb5cd0c959ffa0fbf191634a8aa7d