Change ReportExporter::ExportReport() to use a ReportRowIterator.
Previously ExportReport() took a vector of rows, used
ReportSerializer::SerializeReport() to serialize the rows to
a string, and then wrapped the string in an istringstream that
it passed to the GCS upload utility. Now ExportReport() takes
a RowIterator and constructs a ReportStream that it
passes to the GCS upload utility.

This is most of the aparatus necessary to export a report without
demanding that the whole report fits in memory. But we are still
not there because we don't have any implementation of a
ReportRowIterator other than a ReportVectorRowIterator. So currently
ReportGenerator is still generating reports as a vector of rows and
then wrapping the vector in a ReportVectorRowIterator.

Change-Id: I2f7bb28c52e073340ce91e8f215a45004a6c298b