ReportStream: Set the failbit and the badbit upon error.
Setting the failbit and the badbit on a std::istream is the only way
of communicating to a consumer of the istream that an error
occurred. We need to do this so that the third_party code
that reads from a ReportStream and uploads the report to GCS
knows that a problem occurred.

Change-Id: Id3cef46de4e1eed1b37de2a19eee0c9a13131443