Introduces the ability to not store a generated report in the ReportStore.
We add the parameter in_store in various places. This parameter indicates
that a generated report should be stored in the ReportStore. It will
be useful to be able to generate a report soley for the purpose of
exporting it to an external system without wanting to store the rows
of the report in Bigtable. One example where we will want this is
with RAW_DUMP reports which we will introduce in a later CL.

Change-Id: I4ecae643a2a26faac750ec3a6abb3951e9d8663b