Adds new report type: RAW_DUMP.
A RAW_DUMP report is
a raw dump of the input observations. No aggregation, analysis or
differentially private release will be performed. The report rows will
not be stored in the ReportStore so they may not be retrieved later
via a GetReport() operation. At report generation time the raw dump will
be serialized and exported to an external system if one or more
ReportExportConfigs are specified in the ReportConfig. The external system
may then be used to perform aggregation and analysis.

In this CL we introduce the notion of the RAW_DUMP report and we add logic
to ensure that RAW_DUMP reports are not stored in the ReportStore. But
we don't yet implement the functionality for serializing a RAW_DUMP report.

Change-Id: Ia61fdd08a59ff3df144376f6eebb1bd9ec9b441c