Adds RawDumpReportRowIterator.
This is an implementation of ReportRowIterator that yields the rows of a
RAW_DUMP report. Each yielded report row is essentially a copy of a subset
of a raw unencoded Observation from the Observation Store.
A RawDumpReportRowIterator
wraps a particular query of the Observation store and will incrementally
fetch additional pages of results for that query from the Observation Store
as it yields additional rows.

Change-Id: Idf67a04822cf713dde3d24f460d1de5ffad1705b