Removes obsolete code from analyzer_service.cc
The code being removed was place-holder code for implementing TLS
in the Analyzer Service. We are deleting it for two reasons.
First we are not currently using TLS on the Analyzer Service
because it is not exposed to the internet. Second, if we
did decide to implement TLS on the Analyzer we would copy the
methodology from the ReportMaster on which TLS is enabled.

Change-Id: I27f647c7234bea866d6fd07e28b72db7c2a99743