Add bucketing and distributions in configuration and observations.
In observations.proto, we add the ability to specify a bucket
distribution (a bundle of bucketed numeric observations) as a value.

In metrics.proto, we add the ability to specify a bucketing
configuration for metric parts. It serves as a way to interpret
BucketDistribution observation parts and to bucket observations for
histogram report generation.

Change-Id: Ia5172b3c53a977dfc1e906e0dc070b708a61c3da