Replace bigtable instance name with bigtable instance id.
Apparently a Bigtable instance can have both a name and an ID.
Previously I had never made one where the name and the ID were
different and so I never noticed. But now I have made one
where the name and ID are different and I realize that we
should be using the ID and not the name.

Change-Id: If52d039da4d71a1fcd4718de6a02494f78ae479a