Removes the static IPs from cluster.json.
The static IP listed for the ReportMaster was incorrect and that was
causing cobaltb.py
"deploy start --job=report-master --production_dir=<dir>"
to fail for me.

Some time ago azani@ added a feature where we query the cluster for
the static IP addresses if they are not explicitly specified. Given
that feature there is no reason to even have the static IP addresses
in cluster.json and we should remove that feature.

Change-Id: Ie373ceeb89da920bc5b23847e4b5871095fb9f4f