Adds retries to Bigtable Reads.
Previously we used retry logic for Bigtable writes but not for Bigtable reads.
This CL implements retry logic for reads too.

This is important as it appears to be a common pattern that the first RPC
to Bigtable after the connection has gotten "cold" somehow always returns
a transient failure. In my experimentation this retry logic always succeeds
after one retry.

The symptom I noticed that I am trying to fix is that the ReportScheduler will
frequently skip running a report because it fails to read the metadata for
that report and so it cannot determine if there is already an instance
of the report running. I have observed that this fixes that problem.

Change-Id: I038f31eda7471434bebd2ad21a4fdc9aede529f9