When there are no Observations for a report it should not be an error.
Previously we were returning an error status when there are no Observations
for a report. Here we change to LOG(INFO) and return OK.

The previous behavior is problematic because the returned error causes
the report run to be recorded as an error. This will cause the empty
report to be run every day for 30 days and this spams the logs. With
the new behavior the report will be run only for 3 days after which it
exits the "finalization period" and will not be run any more. This is
the intended behavior of the finalization period.

Change-Id: I4372636ae5b2f7f64a16edd6a6ec05eeaaf8b5fa