ReportExporter: Perform the "ping" operation prior to every upload operation.
The purpose of the ping operation is to get around a bug in 3rd party
code in which when the oauth handshake is performed in response
to a 401 UNAUTHORIZED response, then the content input stream is
not reset. Since the "ping" request uses an HTTP get without a body
there is no content input stream.

Previously we only performed the ping once in order to set-up the
OAuth token. However this is not sufficient as the token expires and
needs to be refreshed and the same problem occurs during refresh. So
now we perform the ping before every upload.

In a separate CL we also fix the bug in the third_party code.

Change-Id: I0db8638beb39990a02517bd4d93fc2d76b7b259a