ReportInputStream: Support a "reset" operation.
When exporting a raw dump report, it may be necessary to reset the
input stream to the beginning if the initial POST request
fails. A typical example is tht the OAuth token has expired. In this
case the server returns a 401 UNAUTHORIZED and then the client
refreshes the OAUTH token and the performs the POST a second time.
It needs to reset the input stream to the beginning.

Change-Id: I889498c8f14ac6532a5facd0fa1ef3fde5b116c1