Try the end-to-end test three times before calling it a failure.
In CI the end-to-end test is flaky. It fails about 1/14 of the time.

The flakiness is inherent in the small non-determinism of a complicated
multi-process test that uses gRPC, so we don't expect to be able to
eliminate it. Our soltion is to give the test several chances to pass
before calling it a failure, thereby greatly reducing the flakiness.

Change-Id: I4e63081c1f8ffd0be1a9c5399a052597467281e8