Add check_config command to cobaltb.py.
The check_config command is intended to make it easier to run the
config_parser tool to check the validity of the cobalt config.
By default, the configuration checked out in the third_party/config
submodule is checked, but a different location can be specified.

We also add an output line that says that the config_parser found no
problem in the configuration directory or repository it was asked to
check.

Change-Id: If2473db1c85ca4f0bd3aad5ff8ed80bf569b7fbb