ProjectContext: Construct from a ClientConfig instead of Metric/EncodingReg
Having introduced a ClientConfig that may be constructed from a serialized
CobaltConfig, now we want to construct a ProjectContext from a ClientConfig.
There are no tests yet. In the next CL we will replace the ASCII proto
in the encoder tests with base64 generated from YAML files and there it
will be natural to test this.

Change-Id: Ia8b2bd22f958b54b821a011392968c3978f4759d