Introduces build macros supporting config  YAML in tests
We are trying to get rid of all uses of ASCII proto in Cobalt because
the use of ASCII proto prevents us from using the proto-lite library
rather than the full proto library.

One place where we currently use ASCII proto is in unit tests. We write
the ASCII proto as string literals within the testing code.

Our new strategy to replace this with a YAML file that lives alongside
the test and contains the configuration for the test. This will be
parsed at build time using Cobalt's config parser which will generate
a binary protocol buffer containing the config, and then generate
a C++ header file containing the definition of a string containing
the base64 encoding of the bytes of the serialized protobuf.

In this CL we are adding a build macro to invoke this.

Since we currently support two build systems, CMake for the stand-alone
Cobalt build and GN for the Garnet build we need to define the macro in
both places.

We also use the new macro once for building encoder_test.cc. Future
CLs will use it more.

CB-102

Change-Id: I83ba0c7257cdb9119fe04010ed2e2d0910aab27f