Add a skeleton for validating configs
This takes a config and returns a StatusOr object that either contains
the validated config, or a status message describing what went wrong.

Additionally, this is added as a "validate" binary that can be
executed by the go config_parser tool. This we can move all of the
validation logic into c++ and still be able to use it in go.

CB-110

Change-Id: I332109273e2361e16ceb6651f18edf18bb3d1eca