Some minor tweaks to get 'deploy production_build' to work for me.
'cobaltb.py deploy production_build' was not working on my machine.
In order to get it working I needed to debug production_util.py
but this was difficult because it started off doing a 10 minute
clean build. So I refactored production_util.py so that doing
the clean build is optional. I also fixed a couple of comments
in Docker files. I also changed an error message and a flag
help in cobaltb.py.

Change-Id: I2ff31014347e0dd0498f2a6a8e5d694eaf3e0390