[encoder] Separate Storage and Shipping logic
This is the first step towards getting non-volatile storage for
cobalt.

CB-58 #comment
We have now separated the storage logic from the shipping logic. This
way, we can swap out the storage logic to use the file system without
having to modify the shipping manager at all. Additionally, it should
be possible for us to run a cobalt encoder daemon that has no
ShippingManager, but instead only writes things to a ObservationStore.

Change-Id: Ibfffccde1ba5926fbbb19cc1c87d032e15972063