Small refactor of CMake build of Clearcut extensions proto library
CB-154

This is an attempt to fix a build flake. The key new ingredient is
the introduction of the statement
"set_source_files_properties(third_party/clearcut/clearcut.pb.h PROPERTIES GENERATED TRUE)"
to the top-level CMakeLists.txt file. But we also took the oppurtunity to
do a small refactor. We moved the build of the Clearcut extensions proto to
its own target because it is only needed in one place. Also we eliminated the
Go build of it which seems to not be used.

Change-Id: Ic6bd93d8871b79e9a24e40bbc3933e42d5d92392