Implements an automatic mode in the Cobalt test_app.
This allows the following to work:

./cobaltb.py start test_app -automatic

This causes the test_app to start in an automatic, non-interactive mode.
We could make the automation configurable in the future if we want, but
for now it is hard coded to use metric 1, encoding_config 1,
and have 100 clients send "www.google.com" once per second.

(My motivation for this CL is I am using it to debug the memory leak.)
CB-155

Change-Id: I1a702700576956586ca153cfb842eaea6835169a