Disable broken test and fix uninitialized variable bug.
There were two bugs in ShippingManagerTest::EnsureObservationsNotSentToClearcut
that were non-deterministically cancelling each other causing the test
to pass most of the time.

(1) The member variable http_client_ was not initialized to null so its
value was non-determined and depended on which tests ran prior to this test.

(2) The logic in EnsureObservationsNotSentToClearcut is wrong because,
when kDefaultMetricId is used then http_client_ is never even initialized
so it doesn't make sense to check its properties.

Change-Id: I874c8db1333a921abc20f10cc5833fe5d374a8a0